Refactor TA review modal
