+++ src/components/tabs/admin/TARow.tsx
@@ -1,20 +1,18 @@
 import React from "react";
 import RemoveIcon from "@material-ui/icons/Remove";
+import { RealTA } from "../../../classes/TA";
 
-const TARow = ({ ta, fetchTAData }) => {
+const TARow = ({ ta, fetchTAData }: { ta: RealTA; fetchTAData: Function }) => {
   const handleDeleteTA = async () => {
     try {
       console.log("deleting " + ta.email);
       // make api call to delete TA
-      fetch(
-        "https://winter2022-comp307-group8.cs.mcgill.ca/ta/delete/" + ta.email,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        }
-      );
+      fetch("https://winter2022-comp307-group8.cs.mcgill.ca/ta/delete/" + ta.email, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
       setTimeout(() => {
         fetchTAData();
       }, 500);
@@ -31,14 +29,12 @@ const TARow = ({ ta, fetchTAData }) => {
           <RemoveIcon />
         </button>
       </td>
-      <td className="column2" style={{ textTransform: "capitalize" }}>
-        {(ta.firstName || ta.backupFirstName) +
-          " " +
-          (ta.lastName || ta.backupLastName)}
+      <td className="column1" style={{ textTransform: "capitalize" }}>
+        {ta.firstName + " " + ta.lastName}
       </td>
-      <td className="column3">{ta.performance}</td>
-      <td className="column4">{ta.currentCourses}</td>
-      <td className="column5">{ta.previousCourses}</td>
+      <td className="column2">INSERT PERFORMANCE</td>
+      <td className="column3">INSERT CURRENT COURSES</td>
+      <td className="column4">INSERT PREVIOUS COURSES</td>
     </tr>
   );
 };
